Splitted wordcloud series into type definition files.
